apply review comments
